amount: Move verification code to submodule
Code move only, no other changes.
